M3_Python_API_Assignments/cs2_BookBuddy_App/bookbuddy/requirements.txt
flask
flask_sqlalchemy
pyjwt